requirements.txt
Flask
Flask-Bootstrap
Flask-WTF
waitress
WTForms
flask_sqlalchemy
flask_login
flask_bcrypt
Pillow
Flask-Migrate
Flask-Script